Migrate to react-native-track-player and play audio even in the background
