+++ index.js
@@ -1,8 +1,10 @@
 import 'react-native-gesture-handler';
 import 'react-native-console-time-polyfill';
 import { AppRegistry } from 'react-native';
+import TrackPlayer from 'react-native-track-player';
 
 import { name as appName, share as shareName } from './app.json';
+import { playbackService } from './app/containers/message/Components/Audio/services';
 
 if (__DEV__) {
 	require('./app/ReactotronConfig');
@@ -20,6 +22,8 @@ if (__DEV__) {
 
 AppRegistry.registerComponent(appName, () => require('./app/index').default);
 AppRegistry.registerComponent(shareName, () => require('./app/share').default);
+
+TrackPlayer.registerPlaybackService(() => playbackService);
 
 // For storybook, comment everything above and uncomment below
 // import 'react-native-gesture-handler';
