+++ package.json
@@ -42,7 +42,6 @@
 		"@react-native-community/masked-view": "0.1.11",
 		"@react-native-community/netinfo": "6.0.0",
 		"@react-native-community/picker": "^1.8.1",
-		"@react-native-community/slider": "4.2.2",
 		"@react-native-cookies/cookies": "6.2.1",
 		"@react-native-firebase/analytics": "^14.11.0",
 		"@react-native-firebase/app": "^14.11.0",
